+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## [1.6.1] - 28-02-2018
+### Updates
+- New format specifier `%session_id`
+- Client integrity tasks are paused for specific clients being processed instead of all clients
+- Unknown client trying to use known logger using valid access code is no longer allowed
+
 ## [1.6.0] - 27-02-2018
 ### Updates
 - Separate log dispatchers and logging queue by clients #96